Test Usuario

+++ ProyectoMolde/ProyectoMolde/ControlUsuarios/Entity/Model/PersonasViewModel.cs
@@ -18,11 +18,11 @@ namespace ControlUsuarios.Entity.Model
         public DateTime fechaExpedicionCedula { get; set; }
         public int usuarioId { get; set; }
         public int? barrioId { get; set; }
-        public int estatura { get; set; }
-        public int peso { get; set; }
+        public decimal estatura { get; set; }
+        public decimal peso { get; set; }
         public int estadoCivilId { get; set; }
-        public int telefonoFijo { get; set; }
-        public int telefonoCelular { get; set; }
+        public long telefonoFijo { get; set; }
+        public long telefonoCelular { get; set; }
         public string numeroDocumento { get; set; }
         public string primerNombre { get; set; }
         public string segundoNombre { get; set; }
