+++ ProyectoMolde/ProyectoMolde/ControlUsuarios/Entity/Model/UsuariosViewModel.cs
@@ -79,12 +79,12 @@ namespace ControlUsuarios.Entity.Model
                 return string.Format("[{0}] [{1}] [{2}]", nombreDepartamentoBarrio, nombreMunicipioBarrio, nombreoBarrio);
             }
         }
-        public int? estatura { get; set; }
-        public int? peso { get; set; }
+        public decimal? estatura { get; set; }
+        public decimal? peso { get; set; }
         public int? estadoCivilId { get; set; }
         public string nombreEstadoCivil { get; set; }
-        public int? telefonoFijo { get; set; }
-        public int? telefonoCelular { get; set; }
+        public long? telefonoFijo { get; set; }
+        public long? telefonoCelular { get; set; }
         public string numeroDocumento { get; set; }
         public string primerNombre { get; set; }
         public string segundoNombre { get; set; }
